feat: create plugin path alongside repo cache
