+++ hipcheck/src/session/session.rs
@@ -162,9 +162,11 @@ impl Session {
 		// Check if a currently unsuporrted policy file was provided
 		// TODO: Remove this error once policy files are supported
 		if policy_path.is_some() {
-			return Err(hc_error!("Policy files are not supported by Hipcheck at this time."))
+			return Err(hc_error!(
+				"Policy files are not supported by Hipcheck at this time."
+			));
 		}
-		
+
 		let (config, config_dir, data_dir, hc_github_token) =
 			match load_config_and_data(config_path.as_deref(), data_path.as_deref()) {
 				Ok(results) => results,
@@ -248,7 +250,7 @@ fn load_config_and_data(
 	phase.inc();
 	// Set the spinner phase to tick constantly, 10 times a second.
 	phase.enable_steady_tick(Duration::from_millis(100));
-	
+
 	// Resolve the path to the config file.
 	let valid_config_path = config_path
 	   .ok_or_else(|| hc_error!("Failed to load configuration. Please make sure the path set by the hc_config env variable exists."))?;
